Refactor draw module (#435)
* refact: draw module

* style: black and isort
